LECTURESIGHT-12: added load/save to UI
git-svn-id: http://localhost/repos/opencast/LECTURESIGHT/trunk@14192 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ src/modules/lecturesight-gui-impl/src/main/java/cv/lecturesight/gui/impl/ConfigEditorPanel.java
@@ -196,6 +196,11 @@ public class ConfigEditorPanel extends javax.swing.JPanel {
     JFileChooser chooser = new JFileChooser();
     if (chooser.showSaveDialog(this) == JFileChooser.APPROVE_OPTION) {
       File file = chooser.getSelectedFile();
+      if (file.exists() && JOptionPane.NO_OPTION == JOptionPane.showConfirmDialog(
+              this, "The file " + file.getName() + " already exists. Do you want to replace it?", 
+              "Replace file", JOptionPane.YES_NO_OPTION)) {
+        return;
+      }
       try {
         config.saveSystemConfiguration(new FileOutputStream(file));
       } catch (Exception e) {
